improve validation map
mudei o validation maps, agora ele checa se o mapa so contem caracteres validos, como a direção do player e os 0 e 1
