+++ maps/invalid.map.cub
@@ -7,15 +7,15 @@ F 220,100,0
 C 225,30,0
 
 111111111111111111111111111111111
-1S0000000011000000000000100000001
+100000000011000000000000100000001
 100000000011000000000000100000001
 110000000000000000000000100000001
 111111111110000000000000111100001
-010000000000000000000000000000001
+0100000000000000000000N0000000001
 100000000000000000000000000000001
 100000001100000000000000000000001
 100000001100000000000000000000001
-100000000000000011000000100000001
+100000000000000011000S00100000001
 100000000000000011010100100000001
 110000011111111111111111100000001
 11110111 1110001 1011110100000011
