+++ src/utils/flood_fill.c
@@ -1,0 +1,37 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   flood_fill.c                                       :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: aldantas <aldantas@student.42.rio>         +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2024/07/03 21:47:46 by aldantas          #+#    #+#             */
+/*   Updated: 2024/07/03 21:49:55 by aldantas         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#include "../../headers/cube.h"
+
+static int	is_valid(int x, int y, char **map)
+{
+	if (x >= 0 && map[x] && y >= 0 && map[x][y] && (map[x][y] == '0'
+		|| map[x][y] == ' '))
+		return (1);
+	return (0);
+}
+
+int	flood_fill(int x, int y, char **map)
+{
+	if (!is_valid(x, y, map) || map[x][y] == 'F')
+		return (0);
+	map[x][y] = 'F';
+	if (flood_fill(x + 1, y, map))
+		return (1);
+	if (flood_fill(x - 1, y, map))
+		return (1);
+	if (flood_fill(x, y + 1, map))
+		return (1);
+	if (flood_fill(x, y - 1, map))
+		return (1);
+	return (0);
+}
